LayoutTests/editing/deleting/forward-delete-empty-table-cell.html

<!DOCTYPE html>
<html>
<head>
<title>Forward deleting in an empty table cell</title> 
<script src="../../resources/dump-as-markup.js"></script>
</head> 
<body>
<div id="edit" contentEditable="true">
<table border="1">
<tbody>
<tr><td>1</td><td>2</td></tr>
<tr><td>3</td><td id="fourth"><br></td></tr>
</tbody>
</table>
</div>

<script>
Markup.description("This tests forward deleting in an empty table cell. The cursor shouldn't move.");

var fourthCell = document.getElementById("fourth");

var range = document.createRange();
range.selectNodeContents(fourthCell);
window.getSelection().addRange(range);

document.execCommand("ForwardDelete");

Markup.dump(document.getElementById("edit"));
</script>
</body>
</html>
